1.4.1: new settings and bug fixing

+++ tools/doneMSG.js
@@ -2,14 +2,14 @@
  * Set the message as a confirm message after an action properly executed
  * @param {Discord.message} msg The message constructor
  * @param {Emoji} emoji An emoji to react
+ * @param {Array} otherMessages Other messages to delete
  * @returns {Boolean} false if error, true if done
  */
-module.exports.doneMsg = async (msg, vars, emoji) => {
-    let channel = vars.client.channels.cache.get(msg.channel.id)
+module.exports.doneMsg = async (msg, vars, emoji, otherMessages) => {
+    const { channel } = msg
     if(!channel) return false
 
     let message = channel.messages.cache.get(msg.id)
-    if(!message || !message.deletable) return false
 
     await message.react(emoji)
 
@@ -32,7 +32,19 @@ module.exports.doneMsg = async (msg, vars, emoji) => {
 
     function deleteMsg() {
         try {
-            message.delete()
+            if(message && message.deletable) message.delete()
+
+            if(otherMessages) {
+                for(let msg of otherMessages) {
+                    if(msg && msg.deletable) {
+                        try {
+                            msg.delete()
+                        } catch (err) {
+                            vars.log(err)
+                        }
+                    }
+                }
+            }
             return true
         } catch (err) {
             vars.log(err)
